IAnalogInputResponseParser returns int and not double

+++ IPX800cs/ActionsExecutors/IGetAnalogInputExecutor.cs
@@ -4,6 +4,6 @@ namespace IPX800cs.ActionsExecutors
 {
     public interface IGetAnalogInputExecutor
     {
-        double Execute(Input input);
+        int Execute(Input input);
     }
 }
